requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url https://nexus.plarin.net/repository/pypi/simple
--extra-index-url https://pypi.python.org/simple

amqp==5.0.6               # via kombu
asgi-lifespan==1.0.0      # via -r requirements.in
attrs==20.3.0             # via automat, pytest, service-identity, twisted
autobahn==21.3.1          # via python-binance
automat==20.2.0           # via twisted
billiard==3.6.4.0         # via celery
celery==5.0.5             # via -r requirements.in
certifi==2020.12.5        # via httpx, python-binance, requests
cffi==1.14.5              # via cryptography
chardet==4.0.0            # via requests
click-didyoumean==0.0.3   # via celery
click-plugins==1.1.1      # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-plugins, click-repl, uvicorn
constantly==15.1.0        # via twisted
cryptography==3.4.7       # via autobahn, pyopenssl, python-binance, service-identity
dateparser==1.0.0         # via python-binance
fastapi-contrib[all]==0.2.10  # via -r requirements.in
fastapi-utils==0.2.1      # via -r requirements.in
fastapi==0.63.0           # via -r requirements.in, fastapi-contrib, fastapi-utils
greenlet==1.0.0           # via sqlalchemy
h11==0.12.0               # via httpcore, uvicorn
httpcore==0.12.3          # via httpx
httpx==0.16.1             # via pytest-httpx
hyperlink==21.0.0         # via autobahn, twisted
idna==2.10                # via hyperlink, requests, rfc3986
incremental==21.3.0       # via twisted
iniconfig==1.1.1          # via pytest
jaeger-client==4.4.0      # via fastapi-contrib
kombu==5.0.2              # via celery
loguru==0.5.3             # via -r requirements.in
motor==2.3.1              # via -r requirements.in, fastapi-contrib
opentracing==2.4.0        # via fastapi-contrib, jaeger-client
orjson==3.5.0             # via -r requirements.in
packaging==20.9           # via pytest
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.18    # via click-repl
py==1.10.0                # via pytest
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycparser==2.20           # via cffi
pydantic[dotenv]==1.8.1   # via -r requirements.in, fastapi, fastapi-utils
pymongo==3.11.3           # via motor
pyopenssl==20.0.1         # via python-binance
pyparsing==2.4.7          # via packaging
pytest-asyncio==0.14.0    # via -r requirements.in
pytest==6.2.3             # via pytest-asyncio, pytest-httpx
pytest_httpx==0.10.0      # via -r requirements.in
python-binance==0.7.10    # via -r requirements.in
python-dateutil==2.8.1    # via dateparser
python-dotenv==0.17.0     # via pydantic
python-multipart==0.0.5   # via -r requirements.in
pytz==2021.1              # via celery, dateparser, fastapi-contrib, tzlocal
regex==2021.4.4           # via dateparser
requests==2.25.1          # via python-binance
rfc3986[idna2008]==1.4.0  # via httpx
service-identity==18.1.0  # via python-binance
six==1.15.0               # via automat, click-repl, pyopenssl, python-binance, python-dateutil, python-multipart, thrift
sniffio==1.2.0            # via asgi-lifespan, httpcore, httpx
sqlalchemy==1.4.6         # via fastapi-utils
starlette==0.13.6         # via fastapi
threadloop==1.0.2         # via jaeger-client
thrift==0.13.0            # via jaeger-client
toml==0.10.2              # via pytest
tornado==6.1              # via jaeger-client, threadloop
twisted==21.2.0           # via python-binance
txaio==21.2.1             # via autobahn
typing-extensions==3.7.4.3  # via pydantic
tzlocal==2.1              # via dateparser
ujson==1.35               # via fastapi-contrib, python-binance
urllib3==1.26.4           # via python-binance, requests
uvicorn==0.13.4           # via -r requirements.in
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
zope.interface==5.3.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
